controlcenter/src/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.34.2
httpx==0.28.1
pyyaml==6.0.2
